Add basic tests for benchmark runners (#85)
* feat: add config argument to all benchmarks

* tests: add first render benchmark tests

* tests: add missing return value

* tests: add deep tree tests

In addition to the deep tree test this patch also adds new config params
to the benchmark component to make it testable. The bootstrap method can
be changed and the benchmark can collect samples if RunService is
unavailable, as it is in the test environment.

* tests: add wide tree test

* fix: off by one error when returning samples

* tests: add triangle test

* tests: no need to skip samples here

All the empty samples are already removed by the benchmark component
itself.

* ci: increase min coverage to 80% again

* refactor: remove useless check before unbind

* Update .github/workflows/test.yml

* Update .github/workflows/test.yml

* tests: address PR comments

* tests: add error boundary to benchmarks

This is to prevent tests from never stopping if a child component throws
an error.

* Update .github/workflows/test.yml